Move assets precompiles tests to tests module
